requirements.txt
streamlit
streamlit_authenticator
pyyaml

